LayoutTests/platform/win/media/controls-drag-timebar-expected.txt

Test that drag the timebar thumb causes seeks.

RUN(video.autoplay = true)
RUN(video.src = 'content/test.mp4')
EVENT(playing)
EVENT(seeked)
Time: 2.3
EVENT(seeked)
Time: 2.8
EVENT(seeked)
Time: 3.4
END OF TEST
